Reorganize the server code.
