+++ .rustfmt.toml
@@ -1,3 +1,4 @@
-group_imports = "StdExternalCrate"
+comment_width = 100
+format_code_in_doc_comments = true
 imports_granularity = "Crate"
-use_field_init_shorthand = true
+wrap_comments = true
