+++ rust-toolchain
@@ -1,1 +1,1 @@
-stable-2024-08-08
+nightly-2024-08-29
