+++ server/src/consts.rs
@@ -1,0 +1,5 @@
+pub const MAX_FRAME_SIZE: usize = 64 * 1024;
+pub const PROD_KEY_FILE: &str = "/etc/ssl/certs/key.pem";
+pub const PROD_CERT_FILE: &str = "/etc/ssl/certs/cert.pem";
+pub const DEV_KEY_FILE: &str = "certs/key.pem";
+pub const DEV_CERT_FILE: &str = "certs/cert.pem";
