+++ server/src/lib.rs
@@ -1,4 +1,8 @@
+mod actor;
+mod attachment;
+mod consts;
 mod error;
+mod handler;
 mod message;
 mod server;
 mod session;
@@ -8,3 +12,5 @@ pub use message::{
     ChatMessage, ClientMetadata, FileReassembler, JoinRoom, LeaveRoom, ListRooms, SendFile,
     SendMessage, WsChatServer, WsChatSession,
 };
+
+pub use consts::{DEV_CERT_FILE, DEV_KEY_FILE, MAX_FRAME_SIZE, PROD_CERT_FILE, PROD_KEY_FILE};
