+++ server/src/server.rs
@@ -1,16 +1,11 @@
 use std::collections::HashMap;
 
 use actix::prelude::*;
-use actix_broker::BrokerSubscribe;
-use base64::{engine::general_purpose, Engine as _};
 
-use crate::message::{
-    ChatMessage, Client, ClientMetadata, JoinRoom, LeaveRoom, ListRooms, Room, SendFile,
-    SendMessage, WsChatServer, WsChatSession,
-};
+use crate::message::{ChatMessage, Client, ClientMetadata, Room, WsChatServer};
 
 impl WsChatServer {
-    fn take_room(&mut self, session_id: &str, room_name: &str) -> Option<Room> {
+    pub fn take_room(&mut self, session_id: &str, room_name: &str) -> Option<Room> {
         log::debug!("Getting room: {}", room_name);
         let session_id = self.rooms.get_mut(session_id)?;
         let room = session_id.get_mut(room_name)?;
@@ -89,114 +84,6 @@ impl WsChatServer {
         Some(())
     }
 
-    pub fn send_chat_attachment(
-        &mut self,
-        session_id: &str,
-        room_name: &str,
-        file_name: &str,
-        mime_type: &str,
-        file_data: Vec<u8>,
-        src: usize,
-    ) -> Option<()> {
-        let mut room = self.take_room(session_id, room_name)?;
-
-        for (id, client) in room.drain() {
-            if id == src {
-                client
-                    .recipient
-                    .try_send(ChatMessage(format!(
-                        "[SystemAck]: File '{}' sent successfully.",
-                        file_name
-                    )))
-                    .ok();
-            } else {
-                log::debug!(
-                    "Sending file {} to client {} in room {}",
-                    file_name,
-                    id,
-                    room_name
-                );
-
-                if client
-                    .recipient
-                    .try_send(
-                        ChatMessage(format!(
-                            "[SystemFile]:{}:{}:{}",
-                            file_name,
-                            mime_type,
-                            general_purpose::STANDARD.encode(&file_data)
-                        ))
-                        .to_owned(),
-                    )
-                    .is_ok()
-                {
-                    self.add_client_to_room(
-                        session_id,
-                        room_name,
-                        Some(id),
-                        client.recipient,
-                        client.name,
-                    );
-                }
-            }
-        }
-
-        Some(())
-    }
-
-    #[allow(dead_code)]
-    // TODO: Implement this method
-    fn send_chat_attachment_in_chunks(
-        &mut self,
-        session_id: &str,
-        room_name: &str,
-        file_name: &str,
-        mime_type: &str,
-        file_data: Vec<u8>,
-        _src: usize,
-    ) -> Option<()> {
-        let chunk_size = 64 * 1024;
-        let total_chunks = (file_data.len() as f64 / chunk_size as f64).ceil() as usize;
-
-        let mut room = self.take_room(session_id, room_name)?;
-
-        for (id, client) in room.drain() {
-            for (i, chunk) in file_data.chunks(chunk_size).enumerate() {
-                let encoded_chunk = general_purpose::STANDARD.encode(chunk);
-
-                log::debug!(
-                    "Sending chunk {} of {} for file {} to client {} in room {}",
-                    i + 1,
-                    total_chunks,
-                    file_name,
-                    id,
-                    room_name
-                );
-
-                let chat_message = format!(
-                    "[SystemFileChunk]:{}:{}:{}:{}:{}",
-                    file_name,
-                    mime_type,
-                    i + 1,
-                    total_chunks,
-                    encoded_chunk
-                );
-
-                if client.recipient.try_send(ChatMessage(chat_message)).is_ok() {
-                    self.add_client_to_room(
-                        session_id,
-                        room_name,
-                        Some(id),
-                        client.recipient.clone(),
-                        client.name.clone(),
-                    );
-                }
-            }
-        }
-
-        Some(())
-    }
-
     pub fn broadcast_room_list(&self, session_id: &str) {
         if let Some(users) = self.rooms.get(session_id) {
             let room_list = users.keys().cloned().collect::<Vec<String>>().join(", ");
@@ -238,109 +125,6 @@ impl WsChatServer {
             rooms.retain(|name, room| !room.is_empty() || name == "main");
         });
         self.broadcast_room_list(session_id);
-    }
-}
-
-impl Actor for WsChatServer {
-    type Context = Context<Self>;
-
-    fn started(&mut self, ctx: &mut Self::Context) {
-        self.subscribe_system_async::<LeaveRoom>(ctx);
-        self.subscribe_system_async::<SendMessage>(ctx);
-        self.subscribe_system_async::<SendFile>(ctx);
-    }
-}
-
-impl Handler<SendMessage> for WsChatServer {
-    type Result = ();
-
-    fn handle(&mut self, msg: SendMessage, _ctx: &mut Self::Context) {
-        let SendMessage(session_id, room_name, id, msg) = msg;
-        self.send_chat_message(&session_id, &room_name, &msg, id);
-    }
-}
-
-impl Handler<SendFile> for WsChatServer {
-    type Result = ();
-
-    fn handle(&mut self, msg: SendFile, _ctx: &mut Self::Context) {
-        let SendFile(session_id, room_name, id, file_name, mime_type, file_data) = msg;
-        self.send_chat_attachment(
-            &session_id,
-            &room_name,
-            &file_name,
-            &mime_type,
-            file_data,
-            id,
-        );
-    }
-}
-
-impl Handler<JoinRoom> for WsChatServer {
-    type Result = MessageResult<JoinRoom>;
-
-    fn handle(&mut self, msg: JoinRoom, _ctx: &mut Self::Context) -> Self::Result {
-        let JoinRoom(session_id, room_name, client_name, client) = msg;
-
-        let id =
-            self.add_client_to_room(&session_id, &room_name, None, client, client_name.clone());
-        let join_msg = format!("{} [SystemJoin] {}", client_name, room_name);
-
-        self.send_chat_message(&session_id, &room_name, &join_msg, id);
-        self.broadcast_room_members(&session_id, &room_name);
-        MessageResult(id)
-    }
-}
-
-impl Handler<LeaveRoom> for WsChatServer {
-    type Result = ();
-
-    fn handle(&mut self, msg: LeaveRoom, _ctx: &mut Self::Context) {
-        if let Some(rooms) = self.rooms.get_mut(&msg.0) {
-            if let Some(room) = rooms.get_mut(&msg.1) {
-                room.remove(&msg.2);
-
-                if room.is_empty() && msg.1 != "main" {
-                    self.rooms.remove(&msg.1);
-                }
-
-                self.remove_empty_rooms(&msg.0);
-                self.broadcast_room_list(&msg.0);
-                self.broadcast_room_members(&msg.0, &msg.1);
-
-                log::debug!(
-                    "User {} in {} left room {}. Current rooms: {:?}",
-                    msg.0,
-                    msg.2,
-                    msg.1,
-                    self.rooms.keys().collect::<Vec<_>>()
-                );
-            }
-        }
-    }
-}
-
-impl Handler<ListRooms> for WsChatServer {
-    type Result = MessageResult<ListRooms>;
-
-    fn handle(&mut self, msg: ListRooms, _ctx: &mut Self::Context) -> Self::Result {
-        let ListRooms(session_id) = msg;
-        MessageResult(
-            self.rooms
-                .get(&session_id)
-                .unwrap()
-                .keys()
-                .cloned()
-                .collect(),
-        )
-    }
-}
-
-impl Handler<ChatMessage> for WsChatSession {
-    type Result = ();
-
-    fn handle(&mut self, msg: ChatMessage, ctx: &mut Self::Context) {
-        ctx.text(msg.0);
     }
 }
